# Use memchr to search for characters to escape