@dralley 2023-10-12
I tried this a while back with `jetscii` and didn't have a ton of luck.  I'll take a closer look at this later this week, it looks interesting.

https://github.com/tafia/quick-xml/issues/405

https://github.com/tafia/quick-xml/pull/408

https://github.com/shepmaster/jetscii/issues/54#issuecomment-1176950186