@Dr-Emann 2023-10-12
Yeah, I actually just posted [a PR to jetscii](https://github.com/shepmaster/jetscii/pull/57#issuecomment-1754287303), and found that even for cases that needed multiple memchr calls, memchr seemed to be faster, and based on the fact that the benchmarks were xml-related, came here to try. 😄