@dralley 2023-10-12
Yes, `pcmpestrm` seems to just be a neglected, minimum-effort instruction.  Whether that's because it never really got adopted or whether nobody adopted it because it's "meh", I'm not sure.

I've seen some interesting ideas about using lookup tables and/or avx256 to get really good performance but haven't really investigated further.

https://github.com/shepmaster/jetscii/issues/22