@Dr-Emann 2023-10-12
I was also playing with a bitset lookup table in https://github.com/tafia/quick-xml/compare/master...Dr-Emann:quick-xml:bitset_escape